Rotatie is nu mogelijk op de camera objecten

+++ Sprint 3/ClientServerDemo/ClientGUI/Json Structure/Serializables/Sub Objects/Transform.cs
@@ -4,6 +4,6 @@
     {
         public double[] position { get; set; }
         public double scale { get; set; }
-        public int[] rotation { get; set; }
+        public double[] rotation { get; set; }
     }
 }
